llm_file.py
# llm_file.py

import os
from dotenv import load_dotenv
from langchain_groq import ChatGroq

# Load environment variables from .env file
load_dotenv()
groq_api_key = os.getenv("GROQ_API_KEY")
# Initialize Groq LLM
llm = ChatGroq(
    groq_api_key=groq_api_key,
    model_name="llama3-8b-8192"  # You can change to "mixtral-8x7b" if needed
)

# Export the LLM instance
__all__ = ["llm"]
